xong thanh toán online

+++ TrungTamTheThao/WebApp/Models/tb_Booking.cs
@@ -44,7 +44,7 @@ namespace WebApp.Models
         [NotMapped]
         public int isCoDinh { get; set; }
 
-        public long? Money { get; set; }
+        public double? Money { get; set; }
 
         public virtual tb_Arena tb_Arena { get; set; }
